Chore: change folder structure for shadcn/ui
