+++ components/WriteLetterButton.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
+import { Button } from 'components/ui/button';
 import { useModal } from 'hooks/useModal';
 import { WriteLetterModal } from './WriteLetterModal';
 import { useSession } from 'next-auth/react';
