+++ components/nav/navitems/KakaoLogoutMenuItem.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Separator } from '@/components/ui/separator';
+import { Separator } from 'components/ui/separator';
 
 import React from 'react';
 import { signOut } from 'next-auth/react';
