+++ components/nav/navitems/LinkClipBoardMenuItem.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { ArrowTopRightIcon } from '@radix-ui/react-icons';
-import { Separator } from '@/components/ui/separator';
+import { Separator } from 'components/ui/separator';
 import { toast } from 'sonner';
 import { useSession } from 'next-auth/react';
